requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.32
pydantic==2.9.0
pydantic-settings==2.5.2
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
email-validator==2.2.0
